Adds validation ${varnameExpr} to ensure it complies with dollarvar. Issue #251.
Since the grammar allows ${Expression} and a LiteralNameOrReference is a
valid expression, and a LiteralNameOrReference may contain '-' since
this is only deprecated and not illegal in the lexer/parser, there is a
need for a special check in this case, when the Expression is a single
LiteralNameOrReference. All other expressions are validated separately.