Adds a Type Evaluator. Issue #253
Adds a TypeEvaluator class PPTypeEvaluator that checks for the static
type of an expression. As the puppet language is highly dynamic, it
respons with either a known NUMERIC, LIST, HASH, VOID, etc. or DYNAMIC.
In some cases it is not possible to state what something really is, but
it is possible to state what it is not. As an example, a dynamically
interpolated string such as "qqq${someVar}" can not be numeric.

At this point, the TypeEvaluator is only tested for NUMERIC.