Corrects check for codesmells selector/case. Issue #178. Closes #259
Earlier issue 178 was implemented using a different approach, but this
was clearly wrong when looking at issue 259. They now work the same way.
If a case expression is the same as the switch expression, all other
case expressions are marked as unreachable. If the same case expression
appears more than once, all equivalent expressions are marked as
duplicates. Optionally, if a missing default is considered a problem it
is marked, and lastly if a default that is not last is a problem it is
marked.