Fixes problem when performing prefixed search with fewer segments
If a search is made for 'sett', a match is expected for 'settings::*',
but did not work due to faulty logic that disqualified 'sett' as not
being equal to 'settings'. 

The fix compares all but the last segment in query (e.g. 'sett') using
equals, and the last segment in query is compared with startsWith in the
candidate's corresponding segment. (The old logic compared the last
segments).