Adds configurable save actions that cleans up spaces. Closes #270.
This commit adds support for three save actions that can be set and
saved per project as well as for the entire workspace. The save actions
all deal with whitespaces; trimming each line, replacing non breaking
spaces, formfeeds, etc. into regular space, and ensuring that the last
line is terminated with a linebreak. These are individually
configurable.