Completes "Open Puppet Element". Closes #271.
This completes the implementation of "Open Puppet Element". All elements
that can be opened and are present in the index are presented, and it is
possible to search using *, and ?. Camel case is ignored. Knows about
namespace e.g. a*::b* (all starting with b in all namespaces starting
with a). This also completes the addition of nicer icons for classes,
variables, definitions and types.