Added support for IIndentationInfo and ILineSeparatorInfo and cleaned up
IFormStream had unused methods, and missed support for some common types
of output.
Indentation now handles a configuration of indent string via
IFormattingContext.getIndenationInformation. It is modified to a Char
and a count.
FormStream now outputs correct line separators (from
IFormattingContext).

As a consequence, the handwritten (to be removed) PPSerializer needed to
be updated. (IFormStream and implementation should be moved).