Adds handling of hidden(...) and fixes issues regarding visible WS.
The HiddenTokenSequencerForDom is now superseded by a
HiddenTokenSequencer in the same package. (The superseded class was
also in the wrong folder (under src-gen).

The new HiddenTokenSequencer is mostly a copy of the class with the same
name in the Xtext package. It adds tracking of rule calls to rules with
hidden definitions and the possibility to override the decision if a WS
is implicit or not.