Completes the implementation of ITextFlow.
An ITextFlow is a "box" of text that text, space and line breaks can be
appended to. Sub-interfaces describe additional capabilities; Recording,
WithText, and ITextFlowMetrics. Three concrete implementations are
also added; TextFlow which implements WithText and ITextFlowMetrics (the
result can be obtained as text, and it is measured), TetFlowRecording,
that measures and can replay the content, but does not produce any text,
and finally MeasuredTextFlow that only measures the flow.

ITextFlowMetrics does not extend ITextFlow in anticipation of caching
such metrics or computation of the same metrics from some other data
source than ITextFlow.

The intended use of the specialized classes is to capture and measure
formatting of nested IDomNodes in an optimized fashion when no concrete
text representation is wanted.

This commit also contains tests for the ITextFlow implementations.