Refactors the formatting responsibilities. Introduces ILayoutManager.
Formatting responsability is now divided between:
A feeder that feeds a dom model to a layout manager (stopping when a
layout manager says it has taken over the responsibility).
A top level formatter that obtains the stylesheet, and sets up
collection of output.
An ILayoutManager that is being fed what to format.

The FlowLayout implementation of an ILayoutManager is now a much cleaner
implementation of what used to be in the CSSDomFormatter.