Adds LayoutUtils with common layout/formatting methods.
Class was added with the intent of containing useful layout/formatting
methods. One such method unifyWidthAndAlign is added. (It calculates the
width of nodes with matching grammar element in a subtree, and then
assigns the max width and given alignment to matched nodes.

A utility class is used to not require inheriting a particular super
class to enable use.