Refactors and adds methods to formatting api.
AbstractSemanticFlowLayout renamed to DeclarativeSemanticFlowLayout.
The formatting API now includes calling before/after methods for
composite as this relieves the user from having to take on the
responsibility of sequencing the children when something is needed
around a composite (e.g. being/end indent). The abstract and declarative
implementations now includes handling of this.

The Select class now has a few more operations:
- GrammarSelector can be initialized from an Iterable (e.g. directly use
result from GrammarAccess.findKeywords(...) and the like.
- Convenience methods to get the Important selector (direct and an And
wrapped Important + supplied selector).