Fixes problems with formatting strings (caused by errors in "hidden").
Not 100% sure of the exact cause, but it seems like unasigned trailing
elements in a grammar rule may be processed with the hidden status of
the parent context. The theory is that semantic sequencing is done and
calls finish (which pops the hidden state), syntactic sequencing then
adds the trailing unassigned - but this is unconfirmed. It could just as
well be a bug in sequencing that is triggered by some other sequence.

This commit works around the issue by checking if trailing abstract
elements are parented by a ParserRule that defines hidden tokens, and if
so, it uses that specification as the current while processing the rule.

As ane example in Puppet, a WS was inserted before the closing double
quote of strings if a particular sequence of text expressions
(interpolated into the string) came before the ending double quote.

With this fix, the correct hidden state are used for strings.