Adds auto line wrapping to text flows.
Auto line wrap kicks in when text does not fit on a line and placement
on the next line will make the result less wide. It is possible to
specify an additional wrap indent that is applied to wrapped lines.

The defaults are to use a maxPreferredWidth of 132, and 0 additional
indent.

The max with is specified as a preferred width since lines may be longer
if it is impossible to wrap/fold them.

Current implementation may leave trailing whitespace on a line and this
whitespace may make width wider than max preferred size.