Improves auto line wrap by collecting unbreakable content.
The previous implementation made a decision as soon as text was being
output. The effect was that if a sequence of unbreakable text was
appended, the append of text that would cause overflow did not trigger a
break since it was preceded by an unbreakable append.

This is now changed - all text between space(0-n)/break(0-n) is
collected, and emitted when a space/break is seen. All measures are up
to date, but may as an effect change if more output is processed.