Adds 'verbatim' output to ITextFlow.
Verbatim text flow output must be supported for "preserve whitespace
mode" since indents should not be inserted, and no automatic wrapping
should take place. 

It should be possible to enter/leave verbatim mode as certain parts
may need formatting, but not others. 

A boolean flag was added to two API methods - appendText(CharSequence,
boolean verbatim), and appendBreaks(int count, boolean verbatim), where
the forms without a boolean is taked as "false". 

In verbatim mode, no indentation is added, and lines are not wrapped.
Since indentation is not emitted it is also not measured.