Links the UI format action to the new formatter.
New PP Formatter is now configured to be used when invoking formatting
in the editor. It uses a slightly heavy handed way of formatting by
re-serializing the semantic model (could probably be built more
efficiently from the node model directly) - more testing needed to see
if it is a performance problem in practice (my guess is that is not
noticeable as this operation is performed manually per file).