Adds functions an style classification of comments acting as breaks.
The DomModelSequencer encodes comments terminated by a line break (as
defined by ILineSeparationInformation) and also if a comment contains
more than a single line. This is done to enable convenient style
classification in styling rules as opposed to having to know the grammar
rules.

IFunctionFactory now has conventince functions that return a
spacing/linebreak specification dynamically depending on if a comment
ends with whitspace or linebreak.