Corrects linebreak and space functions for comments.
When comment is used as a linebreak, it also means that it must accept
that a linebreak is present, or it is impossible to have a comment on
the next line by choice in these postions (i.e. it returns
LineBreaks(0,0,1).

The function for space after comment looked at the wrong node (should
look at predecessor of the following whitespace) - design error. Name of
function changed.