Improves formatting combination of line breaks and comments.
Instead of the previous version to support dynamic functions returning
linebreak and spacing style values this is now part of the default
FlowLayout. A LineBreaks specification now also informs the formatter if
a sucessor comment terminated by a line break is acceptable as a line
break, and if already emitted breaks are accepted as part of the wanted
line break count.

This makes the formatting rules clearer in the normal case, while still
making it possible to completely specify rules for every space/comment
combination.

In this commit, the PP specific formatting tests are broken out to a
separate test class.