Fixes auto correct can destroy content. Closes #321
Auto correct on save processes the content from start to end using the
original string and a regexp matcher find approach. The returned offset
needed to be adjusted for already made replacements. This fix simply
keeps track of diff in offset up to the point of replacement and makes
use of that when making the next replacement.