Improves aggregation of Junitresults with intemediary testsuites.
Prior to this, all testsuites ended up as children of the root suite.
Now they are aggregated into testsuites based on their directory name.

The Junitresult model is also refactored (the testSuites feature moved
from the individual subclasses of the AbstractAggregatedTests - this
reference has now moved to the abstract interface).

This commit also contains a utility Junit test that aggregates a
specific result from a "jenkins stack" - it is not included in the
testsuite, but can be executed separately after having made the result
available under /tmp/jenkins/results. The output appears in the "output"
directory.