Adds hiera functions to all targets. Updates 2.8 references to 3.0. Closes #282
A new 3.0 target has been added, and the previously target known as
2.8 is renamed to 3.0 as there will never be a 2.8 release of puppet.

Internally, all related classes and constants have been changed to
reflect "3.0" instead of "2.8".

Since the string "2.8" may still linger in user's preference stores,
this string is still considered when reading the preferences, but it is
translated to 3.0. When operating on the preferences, the "2.8" is
replaced with "3.0".

This commit also includes improvements in parsing of label and version
of the pptp (version was missing, and labels had an extra space in
them). This because the way they are loaded changed some time ago.
(These labels are only used for documentation when looking at the pptp
XML).

When this commit is used, users that patched their enviroment will need
to remove their patches or will see ambiguity errors for the hiera
functions.