Adds path segment validation in path preferences. Closes #324
Earlier, it was possible to enter a ":" in the field with the effect of
creating multiple entries the next time the preferences were read back
in. This because ":" is the path segment separator. With this change it
is clear that ":" is not allowed, and the value is not accepted.

This change also includes validation of empty segment "//".