Fixes issue with project specific preferences. Closes #320
The project specific settings was stored in project preferences, but not
the flag that turns them on. This resulted in these settings not being
recognized when the project is (re)imported.

Users that update with existing workspace will need to visit their
projects and turn on project specific settings again. The implementation
tries to preserve existing values.

Solution introduces overridden classes from the default Eclipse/Xtext to
deal with preferences in the preferred [sic] way.