Removes workarounds for errors in generated/default sequencing logic.
Prior to 2.3.0 Juno release, the generated sequencers for a language and
the default implementation did not always produce a well balanced
sequence of calls to "enter" and "leave" methods.

This commit removes the work-around. It also contains regenerated PP
logic which triggered the need to also modify how the aux PPTP injectors
are handled as the support in the generated Activator changed.