Modifies serialization to not nullify node model for optional end-tokens
In the previous version, serialization enforced end-commas and
end-semicolons per the style guide. This was done irrespective of an
existing node-model (i.e. "source text"). Unfortunately this means that
it is close to impossible to reconciliate comments (as they are
associated with potentially completely different grammar elements).

This is now corrected - serialization without a node model inserts
separators as per the style guide, and does nothing when a node model is
present.