Corrects winning strategy break vs. space.
The strategy was that break wins if having just a max. With new strategy
break wins if max > 0 && there are breaks (i.e. it only wins if the rule
would be applied). This makes it possible to specify "This whitespace
should be at least one space, but up to 2 linebreaks are ok if present
in the source."