Fixes sequencing issue causing missing implied spaces.
The problem occurred when determining "hidden tokens between 'from',
'to'). The logic only looked at the grammar state at 'to' when
determining if WS before 'to' is hidden or not. 

The change looks at both the from and to hidden specifications, and WS
is implied if one side from/to has hidden(WS).

The problem made it impossible to style these positions when surrounded
by verbatim text.