Adds ResourceAccessScope and ResourceAccess
These clases allows the logic to enter a scope where a Resource (URI)
is bound to the thread. Providers of various instances that are a
function of Resource can thus pick this up and answer with a more
specific instance. 

This is intended to be used to be used to lookup Resource specific
preferences.