Refactors comment handling. Adds w.i.p for formatting preferences.
Comment handling (again) modified and refactored with better separation
of concerns, and reduction of mechanisms (CommentFormattingStrategy
gone - it is just an ILayout).

ITextFlow is now always a measured text flow - all of the
implementations do measure, it was just a hassle to support the notion
of a non measuring text flow.

Adds preference root for formatting with indent size and width. (Not yet
used by formatter). Contains specialized integer field editor with
better error feedback.