Completes General Formatting Preferences. Closes #358 Closes #359
General Formatting Preferences are now implemented and they can be
project specific. (i.e. indentation size, and width).

If "show print margin" is turned on, the general value under
"preferences > ... > Text Editors" is ignored in favor of the puppet
"max preferred width".

The use of the copied ILineSeparatorInformation has been changed to the
class in Xtext 2.3 (it was not available prior to 2.3).