Makes dom node feeder return the "until-node" if a filter is used.
This is useful when feeding content from the dom up to a particular stop
node as it is most likely that the "stop-node" is needed in the
subsequent logic, and also if the stop node was found or not. This
removes the need to search a second time with the same "stop condition".