Fixes indentation issue (indents after wrap point).Closes #362.
The wrapping logic simply made a linebreak after having set the current
"non breakable run" aside, but forgot to also "set-aside" the current
indent, in favor for the indent in effect when the (just set-aside) "non
breakable run". Now the indent at the start of the non
breakable run is remembered and used when a wrap is performed.