Adds a RegionMatch class and tests.
The RegionMatch class is used to compute the relationship between a
CharSequence at a given offset, and a region. Ansers before, after,
starts inside, ends inside, mid inside, or contained. Can also produce
the sub sequences before, inside, and after the region.

This class is needed when a region is formatted and some tokens fall on
the boundaries of the region.