Fixes problem with selective formatting. Closes #371
Selective formatting was never tested with selection starting/ending in
the middle of nodes that have an effect on formatting. There were
several places where this was handled - and done differently. 

The new solution uses one region matcher that computes parts of strings
before, inside, and after the region. Parts outside are not appended to
output and parts inside are appended verbatim (although in the
exceptional case when source is generated all implicit spaces are
appended "non-verbatim" (i.e. can be formatted).