Adds prefix module/project:: to editor tabs for init.pp. Closes #269
It is difficult to tell different init.pp and site.pp files apart when
they are open in editor tabs. This commit adds a modulename:: or
project:: prefix to site.pp and init.pp files.

This is done by looking at the path and using the segment after
"modules" as the name of the module, or if this is missing, the name of
the project.

The prefixing only works for files that are in the workspace - opening
external files is always done without prefix.