Adds Facter 1.6 to the target platform. Issue #167
The interface Facter describes facter facts. The Facter1_6 class is a
concerete implementation for version 1.6. This class is capable of
producing a PPTP model. 

A static copy has been generated and added to the targets in pp.
The handler that syncs the hidden target project now also syncs
facter-*.pptp. (Although at this point there is just support for Facter
1.6, so not much choice what to sync :)

Feature not yet complete, as there are some variables with wildcards
that are found in their base/prefix form (i.e. producing false
positives).