Fixes problem with exceptions when switching target.
Switching targets back and forth often led to logged exceptions and
subsequent builds not functioning. This was caused by timestamp being
cached by Eclipse Resource. This change refreshes the information
in IProoject as well as for the IFile being deleted/replaced.

Probably refreshing too much, but better safe than sorry - the
documentation tries to be clear, but it is confusing if an operation
only modified the workspace notion or not.