Improves token to partition mapping and tab to space mapping. Issues #432 and #430
Whitespace inside string got mapped to the default partition (invisible,
but they were not colored), and they also split the string into multiple
partitions which was not good for selection.

In this commit, also first stab at always mapping tab to space and using
the pp indent size as the tab to spaces map size. Listening for change
of this preference remains to be done.