Makes indent size changes immediately affect tab size. Closes #432.
A rewrite of the Preference change event relay was required to make sure
all events that may have an effect on editor behavior actually reaches
the editor. (It was assumed that when project specific settings were
used, these were the only settings of interest. Not true). 

New implementation relays all events from all stores in the search scope
since a change in any of them may affect the editor for a particular
resource.