Adds preference for automatically closing blocks. Closes #433.
There is now a new preference that controls if blocks are automatically
closed on an enter at the end of a line with opened blocks. This is
useful for those that do not want any automatic help. (This preference
is true by default as it reduces the amount of false negative errors
while typing).