Fixes false negative "unknown variable" outside of path. Closes #441
The problem was caused by search result for "outside" path was discarded
as the search continued to search with a widened scope. This change also
accumulates all found outside.

It may still be argued if a reference to a variable in the same file is
"found outside" or not. But now these are at least warnings (and the
path problem needs to be addressed by the user in any case.