Adds optional alignment of assignment expressions. Closes #442
Assignments using = and += can now (optionally) be aligned when
formatting. If a list contains both = and += the alignment is on the =.
The alignment uses the same clustersize as other alignments (default
20). The optional assignment formatting is off by default, and can be
set per project.

An assignment sequence is only broken by some other type of expression
(whitespace/blank lines and comments are tolerated between assignments).

This allows:

# the foo is a variable that...
$foo      = 10

# the long_bar is also a variable...
$long_bar = 20