Fixes comment reconciliation when L/R association breaks. Closes #452
It is now possible to influence the whitespace sequencing between
comments handled by the comment reconciler. Now, it is also recorded
that a comment sequence has the desire to be "first on line" (as is the
case in issue 452). This mechanism may need to be expanded on in future
situations.