Fixes problem with formatting destroying content. Closes #446
This was caused by a problem in the Xtext framework - it does not ensure
that the document is up to date when performing a modify operation. All
modify operations much be preceded by a readOnly operation (which does
ensure that state is correct).

This affected both save actions, and formatting.