Fixes issue with spaces for tabs not being honored on new. Closes #460
With the fix of #459 where a preference allowing tabs to be input (i.e.
not converted to spaces) was introduced the place where the editor is
initialized was forgotten. The preference was only in effect if it was
changed.

This fix also makes the preference be honoured when editor is opened on
new or existing files.