Fixes NPE when checking attribute names in Collect expr. Closes #467
When linking a collect expression (virtual or exported) and the type is
not known, a null type was passed to the logic that checks the validity
of the attributes. It should simply skip the validation of the
attributes as the unknown type is flagged with an error in this case.