(#471) Fixes java.lang.StackOverflowError when validating long sq string
There is a problem with java regular expressions and using (x|y) as this
can cause very deep recursion. This bit the validator when checking for
unescaped single quotes in a single quoted string (something that can
only happen in model to code transformations as it is not accepted by
the parser in the first place).

This fix corrects this by using a simple non recursive scan of the
string instead of using a regular expression.