Added maven-failsafe-plugin for integration testing with the forge.
This commit adds the maven-failsafe-plugin with pre/post integration
tasks to start and stop the puppet forge server. In order to run the
integration test, a local clone of the puppetlabs/puppet-forge-api
repository must be adjacent to the clone of the geppetto repository.

A database is created and populated from scratch on each integration
test run.