Moves selection of ruby parser to the points where use cases are setup.
This moves the selection of the Jrubyparser from pp.dsl to where the
environment is configured for a particular use case. (This means the
pp.dsl.ui bundle for the IDE case, and to the test suite for the ruby
tests).

Not all that happy about how this is now configured, but since there is
only one working ruby parsing service at the moment it is overkill to
tidy this up further right now. (The "ruby tests" are parser impl
agnostic, but now the test suite depends on the Jrubyparser
implementation). The same is true for the UI bundle; but it is
responsible for much of the build infrastructure configuration and it is
better if the configuration is done adjacent to the rest of the setup.

For the PPStandaloneSetup, the intent is to derive a configuration that
suites a particular use case; such a derived impl should include the
call to setup the ruby parser service factory statically (yes, yuk).