Coherent use of FileFilter when traversing a module.
The forge code used a default regular expression to exclude unwanted
files from the build and from checksum calculations. That approach
was not flexible enough for the publisher since the publisher has
a configurable target directory that needs to be excluded. This commit
changes the use of the default pattern so that we instead use a
FileFilter everywhere. Passing null will result in a default FileFilter
instance which in turn depends on the old default pattern.