Complete rewrite of the Forge service, switching from v1 to v2 API.
Puppetlabs will soon release their v2 RESTful API to the Forge. This
commit ensures that the Geppetto Forge service talks to this new API
instead of the old one. It also refactors large parts of the Forge to
become injectable.

In addition to the refactoring, this commit also adds a new bundle
for sematic versioning.