Moves all embedded tests to separate test bundles.
PDE cannot handle Maven style testing where the tests reside in the
same Eclipse project as the code that it is testing. There's just no
way to get the classpath right for the actual test without also
including it in the final bundle. For this reason, all new tests
related to recent forge publishing development was moved out to
separate test bundles.