Add 3.2.0 pptp to the set of supported puppet target platforms.
This adds a generator of the 3.2.0 pptp file.
The generated file is added to the set of targets.
Utils and handlers updated to map the 3.2 preference to 3.2.0 pptp.

The generated pptp is cut from the puppet master branch while the
external plugins are cut from the 3.0.0 distribution.

The 3.2.0 pptp should be updated once the 3.2.0-rc is out and available
in packaged form (this to ensure that the correct packaged externals are
included).