Improves the styling of lambdas
This adds styling to method calls and lambdas. This applies the
same styling rules to method calls as what is used for functions (i.e.
no spacing before after (). The lambda body is now formatted like
other sequences of statements. (This is better than before but
there needs to also be a compact rendering).