Add formatting of Lambdas.
This adds formatting of lambdas. If a lambda contains a newline
before the closing } it will be rendered on multiple lines. Otherwise
an attempt is made to render it in place (compact form). If the compact
form does not fit on the remainder of the line the lambda is rendered
using non compact form.