Removes cached module choices used when adding dependencies
The list used for picking modules when adding new module dependencies
did not get updated when new modules where added to the workspace
unless the editor was restarted. The list also included dependencies
that had already been added. This commit removes the cache completely
and also ensures that already added dependencies are excluded.

A minor fix for updating the editor title when the module name is
changed is also included.

Closes #294