Adds a listener that closes Modulefile editor when project is closed
The TextEditor tab has a IDocumentProvider that accepts listeners of
type IElementStateListener. This listener is notified when the
IEditorInput somehow changes it state. It can be used for various
purposes such as syncing dirty content or, as in this case, closing
the editor if the resource is no longer accessible.

Closes #303