Major rewrite of the Module metadata editor and metadata validation.
The editor now contains three tabs. 'Overview', 'Dependencies' and
a 'Source' tab. Editing can be made simultaneously in all three
tabs and all changes are kept in sync.

Constant validation of the edtior its content and will display
markers on all three tabs. The 'Overview' and 'Dependencies' tab
will also show an error in the editor title.

The editor can now edit both a Modulefile and a metadata.json. They
are considered different source of input though. When both exists, it
is very likely that the metadata.json file is derived. If the input
is derived, the user will be asked to confirm edits.

The PPModulefileBuilder was renamed to PPModuleMetadataBuilder to
reflect the fact that it now also will verify the metadata.json file.
This happens when the Modulefile is missing and the metadata.json is
not derived.

Closes #508, #509, #510