Improves tracking of dependencies in Metadata Editor and Builder
The new positional parsers for Modulefile and metadata.json will add
file and location to the parsed dependencies. This information is
then used by both markers in the problem view and during editing of
the files.

This commit also takes care of the editor loosing track of which
dependency that was added when its name was changed.

Closes #518