Adds an extra "JSON (derived)" tab and an IEditorMatchingStrategy
The module metadata editor did not convey the fact that the
'metadata.json' is (typically) a derived artifact when a Modulefile
is present. This commit provides a large improvement in that it
will present both files as separate tabs in the same editor. The
additional tab is named 'JSON (derived)' and is a read-only display
of the derived artifact.

An IEditorMatchingStrategy was added to ensure that the same editor
opens regardless of which file (Modulefile or metadata.json) that is
clicked and that only one editor is opened. Clicking on the
'metadata.json' file while the editor is open will simply activate
the 'JSON (derived)' tab. A click on the 'Modulefile' will flip to
the 'Overview' tab in case the 'JSON (derived)' tab is active.

The use-case where no Modulefile exists is still supported. The
'Source' tab was renamed to 'JSON' for consistency and of course,
no fourth tab is shown.

Some improvements were also made to the "new Modulefile" wizard. It
now ensures that a derived 'metadata.json' is present before the
editor opens (or the fourth tab will be missing). The same improvement
was made to the 'New Module Project' wizard.

The changes to the title of the 'Overview' page suggested in issue #516
is also covered in this commit. The same change was made on the
'Dependencies' page.

Closes #508
Closes #514
Closes #516