Ensure a full build on startup when Geppetto version has changed
This commit inserts code into the o.c.g.pp.dsl.ui bundle activator
start method that checks if the current version of the bundle is
equal to a 'last.builder.version' persistent property of the workspace.
If this is not the case, or if the persistent property is nonexistent,
then a full build of the workspace is triggered and a new value is set
for the persistent property.

Closes #515