Fixes problem causing the wrong file to be displayed
Having a Modulefile editor up in one project and then clicking the
metadata.json file in another project simply switch the first editor
to show a metadata.json (the wrong one). This commit adds code to the
matching strategy that checks if the parent container is the same for
the matched files.