Ensures that markers maintained by editor are refreshed after a build
The ModuleMetadataEditor maintains its own set of markers (specific
to the editor). This commit ensures that those markers are updated
when a workspace build has been performed and thus makes them reflect
the same thing as the build markers (well almost. The build cannot see
unsaved changes).

A previously missing dependency will now get its marker removed as
soon as a project is added that fulfill the requirement.

Closes #524