Make Geppetto work with httpcommons 4.1
This "downgrades" Geppetto so that it can play nicely with the
Eclipse IDE with respect to org.apache.httpcommons plug-ins.
Apparently, the ones bundled with the IDE (even in Kepler) has still
not been upgraded from version 4.1. This means that the HttpPatch
and URIBuilder classes are not included and must be faked.