Fix problem with re-added module missing the Puppet Nature
This commit inserts a check to see if the target directory for a new
project is empty. The user is asked to confirm an overwrite when it
isn't. The directory is emptied prior to creating the new project if
the user confirms which means that no old .project file can be
lingering and screw up the creation of a new project.

Closes #528