Replace all OSGi Bundle-Requirement with Import-Package
Bundle-Requirement has a lot of cons since it implies that the
consumer has knowledge of what bundle it is that provides what
packages in the OSGi runtime. In many cases its use leads to
conflicts.

The recommended way is to always use Import-Package instead and
combine that with 'uses' annotations on the exported packages. The
'uses' clause gives the OSGi runtime a hint on how to configure
classloaders in situations where there are several providers of one
package.

Good article on the subject:
http://olegz.wordpress.com/documents-and-articles/osgi-uses/