Limit version of com.google.inject to "[1.3.0,1.4.0)"
The 1.3.0 is Guice official release. The maven runtime bundle delivered
as part of m2e contains 'sisu' whith version 1.4.0 of com.google.inject
but with no com.google.inject.multibindings. This leads to a conflict
if this requirement is missing since the OSGi loader will attempt to
load the latest com.google.inject package.