Use Forge Login setting as the default Module owner
The "New Module" wizard was using the System property "user.name"
as the default for the module owner. This was bad since that name
may contain illegal characters. A better default is to use the
preference "Puppet" -> "Puppet Forge" -> "Module Forge Login".
Especially since the forge requires that the authenticated user is
equal to the module owner anyway when publishing.

The change will have no effect for most users since the preference
was initialized using the system property "user.name" anyway.

Closes #537