Add missing call to super.dispopse()
The ModuleSourcePage extends TextEditor and overrides its dispose()
method. The override failed to call super.dispose() which had the
bad consequence that the associated editor input was never released.
Since it was never released, it's in-memory cache remained effective
and when creating a new document with the same name, the previous
content was presented. This commit adds the needed super.dispose()
call.