Ensure that form message is displayed on first open of editor.
When a module has dependency errors and the Modulefile editor was
opened, the red 'Module has unresolved dependencies' text was not
immediately visible on the overview page. This was because it was
added to the message manager before the page was visible. This
commit traps the pain event on the composite body of the form and
ensures that the message becomes visible on first open.